refactor: align LiveKit adapter with SDK patterns
